+++ .env
@@ -1,4 +1,3 @@
-
 workers=1
 data_dir=temp
 delete_files_on_success=false
