+++ env.go
@@ -1,0 +1,44 @@
+package main
+
+import (
+	"log"
+	"os"
+	"strconv"
+	"strings"
+)
+
+func mustGetenv(key string) string {
+	val, ok := os.LookupEnv(key)
+	if !ok {
+		log.Fatalf("env var must be defined: %s", key)
+	}
+	return val
+}
+
+func getEnvString(key string, fallback string) string {
+	if value, ok := os.LookupEnv(key); ok {
+		return value
+	}
+	return fallback
+}
+
+func getEnvInt(key string, fallback int) int {
+	if value, ok := os.LookupEnv(key); ok {
+		i, _ := strconv.Atoi(value)
+		return i
+	}
+	return fallback
+}
+
+func getEnvBool(key string, fallback bool) bool {
+	if value, ok := os.LookupEnv(key); ok {
+		if strings.ToLower(value) == "true" {
+			return true
+		}
+		if value == "1" {
+			return true
+		}
+		return false
+	}
+	return fallback
+}
