+++ main.go
@@ -8,7 +8,6 @@ import (
 	"os"
 	"os/signal"
 	"path/filepath"
-	"strconv"
 	"strings"
 	"sync"
 	"time"
@@ -97,14 +96,6 @@ func fillJobQueue(stop <-chan struct{}, root string) (<-chan Job, <-chan error) 
 	return jobs, errc
 }
 
-func mustGetenv(key string) string {
-	val, ok := os.LookupEnv(key)
-	if !ok {
-		log.Fatalf("env var must be defined: %s", key)
-	}
-	return val
-}
-
 func mustGetS3Uploader() *S3FileUploader {
 	endpoint := mustGetenv("s3Endpoint")
 	accessKeyID := mustGetenv("s3accessKeyID")
@@ -136,34 +127,6 @@ func mustGetS3Uploader() *S3FileUploader {
 		Session: session,
 		Bucket:  bucket,
 	}
-}
-
-func getEnvString(key string, fallback string) string {
-	if value, ok := os.LookupEnv(key); ok {
-		return value
-	}
-	return fallback
-}
-
-func getEnvInt(key string, fallback int) int {
-	if value, ok := os.LookupEnv(key); ok {
-		i, _ := strconv.Atoi(value)
-		return i
-	}
-	return fallback
-}
-
-func getEnvBool(key string, fallback bool) bool {
-	if value, ok := os.LookupEnv(key); ok {
-		if strings.ToLower(value) == "true" {
-			return true
-		}
-		if value == "1" {
-			return true
-		}
-		return false
-	}
-	return fallback
 }
 
 func main() {
